requirements.txt
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
colorama==0.4.6
dj-database-url==1.0.0
Django==4.1.3
django-bootstrap-modal-forms==2.2.0
django-cors-headers==3.13.0
django-heroku==0.3.1
django-storages==1.13.1
dropbox==11.36.0
gunicorn==20.1.0
idna==3.4
numpy==1.23.4
opencv-python==4.6.0.66
Pillow==9.3.0
ply==3.11
psycopg2==2.9.5
qrcode==7.3.1
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
stone==3.3.1
sweetify==2.3.1
tzdata==2022.2
urllib3==1.26.12
whitenoise==6.2.0
